Register hardcoded hotkey for now
Hardcode to Win-Alt-G until we can get a proper preferences panel up.

Registering the hotkey properly (so that it would show the window
immediately without requiring the window to show at least once) required
moving the window initialization into OnStartup instead of lazy-creating
it the first time it needs to be shown.