Bump to .NET 4.6.2, fix HTTPS image preview.
MediaElement's Source property cannot handle HTTPS images
due to some broken security checks, and is not going to be
fixed by the WPF team anytime soon (per Connect issue), so
work around it in our own insane way by embedding an HTTP
server that will be used as a proxy for HTTPS images for
preview. Once they're added to the library, we use the locally
cached copy as a source, so we needn't worry about wrapping
there.

Fixes #36.