Add a virtualizing wrap panel for image list
As it turns out, if you try to load even a measly 75 images into a
regular wrap panel, the resulting media element cascade eats 3-ish GB
of RAM and chews CPU for 10-15s. This is Not Good. Virtualizing the
wrap panel seems to work pretty well for keeping things snappy, even
if you do have to wait for a bit once elements scroll into view. This
could probably be improved with some smarter scroll handling. :-)